NANGO-677: Allow the websockets path to be configured
# Summary

The websockets path is currently hardcoded to `/` which makes it hard to
firewall or otherwise filter requests to the dashboard (which should be
very secure) from requests to the websockets (which must be open
publicly).

This adds:

- `NANGO_WEBSOCKETS_PATH` to the server's (optional) environment
  variables.  This defaults to `/` for backwards compatibility
- `websocketsPath` as an optional config parameter to the frontend
  `Nango` control. This defaults to `/` for backwards compatibility.

*WARNING*: This PR is not complete and requires work to update the
WebApp to also allow it to use an updated websockets path for the
inline OAuth flow on the `Add New Connection` page.

# Testing

TBC based on completing the above.