Add support for Microsoft tenant-specific endpoints (#1260)
#  Summary

The `microsoft-teams` provider works for shared/saas Microsoft services
that are in the `common` tenant.  This provider is essentially the same
but allows the `tenant` to be provided as a Frontend SDK additional
config parameter so that any tenant can be accessed.

This is required for services that are tenant specific and are not
available in the general `common` tenant as the tenant id can be a
random value.

This has been added as a new provider rather than changing the
existing `microsoft-teams` one so as not to break backwards
compatibility for users of that one.  This new provider also does
not provide any sync configuration as it is unknown which service
will be behind the Azure app registration so there isn't anything
consistent to sync.

See the docs update included with the PR for more details.

# Tests

This provider has been used successfully in local development to
connect to Microsoft Dynamics 365 Finance & Operations and successfully
call those endpoints after authorising using Nango and a custom Azure
App registration (after adding the `Dynamics ERP` API permissions to the
App registration).