frontend: adding logs to debug why Nango.auth() sometimes doesn't resolve
There is a possibility that the frontend doesn't connect to the
websocket server fast enough (aka before the server sends the
notifySuccess/Error). In this case the success/error message is never
send.
This commit adds logging in notifySuccess/Err to test this theory and
confirm if the server successfully sends those websocket messages