fix(activity): reduce time and memory allocated for cleaning the table (#1616)
The query to clean up the table has suddenly been failing for many days now because of timeout. Usually that's a sign the table is too big and the row deleted do not fit in memory anymore, so PG needs to write to disk before deletion making the query much much slower.
Postgres does not allow DELETE LIMIT so the trick is to batch smaller and more often.
LIMIT 5000 can be tweaked, but usually 10K is the advisable maximum.