fix(activity): limit the number of logs displayed (#1665)
Some jobs can contains 100K+ messages which is not playing well with the API nor the UI.
This is a very short term solution to stop crashing our servers (and customers UI when the request do come).

Right now the query is getting "N activity with N messages" which is cumbersome to achieve in Postgres.

I feel like a better solution (without changing too much things) would be to have one call per Activity. That would make the query far more straightforward, with the compromise of making more queries in the UI. Since this fix would have changed a lot the UI logic, and there is redesign on its way, I didn't wanted to go this way.