Do not use -w option when upgrading shared dependency
We ran several times into a situation where shared dependency was not
upgraded properly when publishing a new version.
I (Thomas) suspect the -w option might be buggy and causes some of the
package not to upgrade the shared version properly.
This commit gets rid of the -w option in favor of navigating into the
package folder and running npm install there directly.