fix(cron): delete syncs naive approach (#1776)
## Describe your changes

This cron has proven much more complicated than anticipated. It's not
easy to say to PG `find me what should have been deleted but that is not
deleted". Basically it has to do a seq-scan on all the tables, which is
not good.

- Revert to a naive approach, select every sync that has recently been
deleted and iterate over them.
That means more useless work but at least it should be much faster

- Use advisory lock to prevent other cron to trigger at the same time


This change the fact that if the cron didn't work for a moment it will
not catch up, but since it's impossible I prefer to focus on the best
case scenario until we have a queue to manage jobs.