fix(integration):stripe-app access token refresh on expire (#1780)
## Describe your changes
Nango was finding it hard to refresh `stripe-app` access token upon
expiration. This was because upon exchanging a code for an access token,
there is no `expire_at` or `expire_in` value in the response. However,
according to the
[docs](https://docs.stripe.com/stripe-apps/api-authentication/oauth#refresh-access-token),
the access token expires after an hour, which is equivalent to 3600
seconds. I added that value during the creation of the Stripe app token.
This will help facilitate automatic refresh on access token expiration.