fix: jobs vulnerability
jobs depends on the api package to make request to render api This
package suffers from a vulnerability of one of its dependency,
specifically the lodash.setWith package which is actually deprecated.
There is a PR open for api to use full lodash instead of per method
packages (which are deprecated) but it has not been merged yet.
https://github.com/readmeio/api/pull/859

This commit replaces the api package used to generate a render sdk from
their openapi spec by a home-made RenderAPI class (40 lines of code)