fix(ui): move and pin devDependencies (#1808)
## Describe your changes

Fixes NAN-519

- Move everything to `devDeps`
We install everything the UI as prod dependencies, which is irrelevant
since the build does not require any dependencies.
Also if we go for a unified Docker it will reduce the install size by a
large amount.

- Pin dependencies
Open to debate. Having a caret means you are okay with transparent minor
update, which in NPM ecosystem is very optimistic. Unless it's a
library, in which case you do want a caret, pinning creates a stable
reproductible build.
See recommendation
https://docs.renovatebot.com/dependency-pinning/#so-whats-best

- Removed `jest` as we are not using it