[nan-544] allow optional params to be passed in and used (#1841)
## Describe your changes
See the community reported issue:
https://nango-community.slack.com/archives/C04ABR352H0/p1710132351607969

The user is triggering an action from a sync in an effort to abstract
out logic. This used to work until optional parameters were not allowed
to be passed into a getConnection call with the changes in
https://github.com/NangoHQ/nango/pull/1764/files#diff-f52f0d61cd6802499f5a31aca89d3aec3df278a751065390c02b6bbdb24c0840L337-R341

This update allows optional params to be passed into a getConnection
call. The ideal case is that the caching of a connection call is keyed
by a `providerConfigKey` and `connectionId` combination but I feel that
is an edge case that is out of scope of the immediate user reported
issue and this update resolves their blocking issue.

## Issue ticket number and link
NAN-544

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: