feat: interactive demo (#1801)
## Describe your changes

Fixes NAN-517 NAN-522

The Getting Started page is the main entry point when registering to
Nango and it's well used but do not currently show everything that is
good with Nango. The goal of this revamp is to highlight the main
features, make it more explicit that it's a demo and not really a
getting started, have a little bit more interactivity and simplify also
the sync to only pull from our own controlled repo.

- Refactor the API 
The flow looks like this now:
   - `POST /` to create a row 
- `GET /status` with a deep check on the differents parts to be sure the
progress is accurate if some stuff failed or user has deleted
syncs/provider
   - `POST /status` to update progress 
   - `POST /deploy` to create the sync/action
- `POST /sync-status` to check on sync completion and launch it if it's
hasn't started or paused
   - `POST /action` to trigger the action manually 
 
- Refactor the UI
Split all the components for maintainability. The review before->after
is hardly possible but you should consider that everything has changed.
The Figma (Interactive Demo - WIP) is available to check the design.

 - Removed server side analytics tracking
 It's not useful as we already track in the frontend.

## Screenshot

<img width="1512" alt="Screenshot 2024-03-12 at 11 36 57"
src="https://github.com/NangoHQ/nango/assets/1637651/1d555a7b-0e34-4aa3-9cb3-5a2a2a48bc21">

## How to test

- `git checkout feat/getting-started-revamp`
- Update the .env variables `DEFAULT_GITHUB_CLIENT_ID` and
`DEFAULT_GITHUB_CLIENT_SECRET`
- `npm run dw` `npm run dwa`
- http://localhost:3000/dev/interactive-demo

After you execute the Action please make sure to delete the resulting
Issue in the repo. We will need to clean up regularly this repo, the
Fetch is already filtering on the label but it's just to be clean.