fix(ui): use hooks for meta/env/user (#1844)
## Describe your changes

Fixes NAN-558
Contributes to NAN-516

We were spamming a bit /meta /env /user because we were using straight
http call in the components.
This is tentative to start refactoring those calls using SWR and hooks.
I also need to add stuff to `/meta` for the new interactive demo and it
was a good time to have something stable to do it later.

- Start exporting some types from the Server
We are duplicating types from server to webapp, which creates
discrepancy. Start small by exposing `GetMeta` to see if the forms is
fitting.

- Introduce `useEnvironment` `useMeta` `useUser` 
This should remove the biggest offender that fetch on every pages. This
also reduce the boilerplate needed on some pages, but could be reduced
further by removing some useless setState.