fix(demo): hide after complete (#1849)
## Describe your changes

Fixes NAN-516

- Hide demo after completion
This will only happen after refresh. I was going to animate it but
definitely not worth the time at this stage.

- Change the way we handle 404 and redirection
Because I needed to wait for meta to be ready before making a
redirection (I don't want to redirect to the demo if we already
completed it) I needed to have a hook instead of a function. I split
NotFound, Homepage, and handled the different scenario there.

---

NB: I noticed this bug
https://linear.app/nango/issue/NAN-568/fix-env-in-url-is-not-synced-to-cookie