fix: deleting lots of records (track_deletes=true) (#1878)
I am a few days away from removing the records snapshot table but a
customer is experiencing a sync failure in prod. so here is a quick fix
to unblock the customer.
https://nangohq.slack.com/archives/C06FA6Q90BF/p1710782759958379

Deleting records (track_deletes=true) currently involves copying records
from the snapshot table back into the records table. If there is a lot
of records to copy, the query fails because the number of parameters
goes beyond what's postgres supports, which I believe is 65k
Fixing this issue by inserting in chunks

## Issue ticket number and link

https://linear.app/nango/issue/NAN-603/bug-deleting-too-many-records-track-deletes=true-results-in-an-error

## Tested?
I have reproduce the issue and tested the fix locally