chore: unified dockerfile (#1840)
> [!NOTE] 
> This is a proposal, open to debate.

## Context

Fixes NAN-596

We currently maintain 4 Dockerfiles, and need to build 7 in total. I
feel like it's a lot to maintain. As we grow we will add more images and
more folder in packages which can results in more stuff being build for
nothing and errors.

I would love to reduce this maintenance to 3 images (`staging`, `prod`,
`enterprise`) and down to 1 if possible. Right now we need many versions
also because we pass dedicated env vars to the Webapp build which is
something we could do differently (e.g: serving those via API calls)

## Proposal

- Build one unified Docker image `nangohq/nango` 

It's currently 148MB (compressed)
[hub.docker.com](https://hub.docker.com/repository/docker/nangohq/nango/tags?page=1&ordering=last_updated)
vs `nangohq/nango` is 145MB. So the difference is already minimal enough
that it doesn't really matter.

NB: the image is private rn

NB: I disabled build for `entreprise` for the moment. I feel like it
would be a waste of time and if we manage to get rid of the Env Vars
situation, the single image could serve this scenario.

## Migration

If we agree the migration path is not urgent. We just need to take the
time to do it when we want.
E.g for jobs:
- changing the image to `nangohq/jobs:production` to
`nangohq/nango:prod-<hash>`
- adding a `Docker Command`, e.g: `node packages/jobs/dist/app.js`
- When we deploy, instead of triggering a new deploy on the same tag, we
can trigger a deploy with the new image url

NB: this should also solve our docker caching issue

### Should we keep the other images?

In the long-term I feel like it's not really useful, I think what other
projects are doing is perfect with only 2-3 images:
- project/image_community
- project/image_entreprise
- project/image_private (for us)