fix(cli): glob change of behavior (#1891)
## Describe your changes

Followup of #1874 

When I did this eslint pass, I found undeclared usage of packages. It
was the case for glob, so I had to pick a version. After testing, I
picked the last one.
While `nango cli dev` worked, it turns out it had a slightly different
logic than `nango deploy dev` which hide a change in behavior.

- Share a single unified way to listFiles
- Add tests

## How to test?

- npm run dw
- `node ../nango/packages/cli/dist/index.js dev`
- modify a sync
- `node ../nango/packages/cli/dist/index.js deploy dev`
- check that your file are still there and compiled