fix(utils): parse env with zod (#1919)
## Context

Fixes NAN-662

Currently we are using `process.env` everywhere but it's untyped and
always a pain to deal with `any`. The idea with this function is that
each service can call it when they launch and combine it with their own
set of required env vars.
For example in logs I need Elasticsearch env but I can't make it
globally required. (I'm sure it will be handy for @TBonnin too)
You can find an usage example here:
https://github.com/NangoHQ/nango/blob/1085b3149239da8c2bd050fffc1021e39c34f061/packages/logs/lib/env.ts#L1-L4


The main drawback is that it can't be used for floating constants like
we have in `detection.ts`. I guess **we need to decide** wether it's
fine to have this function run directly or make the detection constants
also an object in a function. I will say it's a bit less convenient for
sure, but with less side-effects
```ts
function initDetection(envs) {
  return { isCloud: envs.NANGO_CLOUD }
}
export const { isCloud } = initDetection(initGlobalEnv());
```

## Changes

- Expose a way to parse process.env with Zod (not exhaustive)