feat: logs storage v2 (#1865)
## Describe your changes

Fixes NAN-580

> [!NOTE]
> This is the implementation only, it's not used anywhere to limit PR
size and possible issues at deploy

- Introduce ElasticSearch in the stack
Locally that means a single node in Docker, and also in tests.
Integration tests are a bit slower to start because of that.

- Create client, model and helpers
This is a "in theory" implementation. It should cover most use cases
(and it's tested) but actual usage might require some fine tuning.

## ElasticSearch details

- One index with everything (Operation and Messages) 
Because we want a unified logs UI. It's a bit like in Datadog traces,
everything is a span, the main Span just don't have parent. It's the
same here, Operation don't have `parentId` otherwise they share the
exact same properties.

- No partitioning
For the moment I'll go for one index, mostly because we don't have that
much data, I don't want to over-engineer this feature. Also Operation
and Messages can spread for many days and it might be too naive to split
by days.
Retention can still easily be achieved by using `deleteByQuery` 

## Usage
> You can read the tests to see more

```ts
const ctx = await getOperationContext(operationPayload);
// == const activityLogId = await createActivityLog();

ctx.info('This is an info');
ctx.success();
// == createActivityLogMessageAndEnd
```