add records package (records migration) (#1933)
## Describe your changes

Adding the `records`package to interact with nango users records. This
package introduces a new table which will be hosted on a different
database.
For now the package is not being used, other than running the db
migration. Next step will be to use this package to double write the
records in both the old table and the new one.

## Issue ticket number and link

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [x] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: