[nan-668] add verification bash script to publish packages (#1930)
## Describe your changes
The goal with this PR is to mimic a publish process of the cli with the
code that is in review to ensure a working cli that is in a remote
source. This is accomplished by building and publishing the packages to
the Github registry and finally using the cli to verify that it works
expected. This will protect against instances where a faulty cli is
published. This will be especially useful to verify that this PR:
https://github.com/NangoHQ/nango/pull/1918 works as expected with
packaging the utils package. Additionally, as we refactor the cli to
move code out of shared and reduce the package size of the cli this will
ensure things still work as expected.

## Issue ticket number and link
NAN-668

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: