fix(webhooks): do not await send (#1977)
## Describe your changes

Fixes NAN-721

> [!NOTE]  
> This is a proposal

When we create a new connection we send a webhook synchronously which
blocks the response until the webhook is sent.
It's debatable wether or not this should be in the critical path. I
guess it's handy to have it sync but it's annoying if the backend
receiving the webhook is down or slow. The ideal fix is to put a queue /
dedicated webhook service

If you think this behavior is wanted/fine, then we can close this PR.

- Do not await webhooks in the critical path