fix(node): use version 20 (#1989)
## Describe your changes

Fixes NAN-732

Node20 is the LTS since october, it's now mostly stable and will provide
new API and perf upgrade.
The codebase "should" stay compatible with 18 until April 2025 (we are
still running the tests in 18 but we are not doing matrix tests with
everything). We have to be careful with new API only available in v20 or
we can also decide to entirely drop support for v18. However I think
it's safer to support 18 for a longtime because of the node-client and
cli.

- Use Node20 by default
- Upgrade related github actions package