Move Result and core Encryption logic to packages/utils (#1999)
## Describe your changes
For convenience reason `records` package was depending on `shared` for
`Result` and `EncryptionManager`. To implement records writing we need
`shared` to depend on `records` to mark records as deleted when
`track_delete = true`, which would create a dependency cycle between
`shared` and `records`.
This PR is moving `Result` and the core of the `Encryption` logic to
utils so `records` package doesn't depend on `shared` anymore. And
anyway no library package should depends on `shared`