fix(release): unify build  (#2026)
## Describe your changes

> [!NOTE]
> This is a proposal

Each package has its own slightly different `npm run build` command. It
is confusing and actually redundant with the global build command
(except for the CLI but hopefully we will be able to remove that at some
point). And we will add more packages so we want a something simple that
works

- Proposing to reduce to one command  `npm run ts-build`
- Remove unnecessary commands