fix(records): transaction not reusing opened connection (#2045)
## Describe your changes

Contributes to NAN-796

There is discrepancy is the new records storage, we have a lot of error
in production related to pool exhaustion which is probably due do the
fact that the transactions were not reusing the connection correctly.
Not entirely sure it's the full fix but maybe

- Fix transaction usage
- Still increase the pool since we have a better DB

```logs
2024-04-25T08:58:41.431Z [ERROR] [records] Failed to upsert records to table records.
Model: ***, Nango Connection ID: ***.
Attempted to insert/update/delete: 100 records
KnexTimeoutError: Knex: Timeout acquiring a connection. The pool is probably full. Are you missing a .transacting(trx) call?
2024-04-25T08:58:41.431Z [ERROR] [PersistController] Failed to save records: Error: Failed to upsert records to table records.
```