[NAN-842] implement unused token refresh (#2094)
## Describe your changes
Adds a cron to refresh stale credentials. Stale credentials in this case
are connections that have not been fetch in the past 7 days. With the
current number of credentials in prod that meet this criteria with a
limit of 500 this will take about 24 days to refresh all of them based
on the cron just running once a day.

Note that this is starting off relatively conservative to track load and
also monitor metrics.

Logic here is intentionally light initially to build upon with what we
spoke about in the meeting the other day (with notifications, ability to
set refresh interval etc.)

## Issue ticket number and link
NAN-842

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is:

---------

Co-authored-by: Thomas Bonnin <233326+TBonnin@users.noreply.github.com>