fix cap logic (#2108)
## Describe your changes
On deploy the cap was being applied but only to disable the sync config
which if a sync was already had no impact. This update fixes the cap
application and adds an event for deploy so we know it isn't an active
cap but rather a passive one. A passive cap in this case means that it
doesn't disrupt any operation but if a new connection is created the
sync won't initiate because it is disabled.

## Issue ticket number and link

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: