fix: latency and type error (#2112)
## Context

Fixes NAN-894
Fixes NAN-893

With the latest deploy we saw an increase of latency that does not show
up in metrics at all (only the global latency increase but no other
metrics is showing the same increase). Which lead me to believe its a
pool acquisition problem. Basically we are doing more queries so we
exhaust the pool more rapidly, the queries are still fast but wait time
is increased.

- Increase default pool size
I could have increase the env var, but I think it reflects our
expectation.

- Track pool usage
It's a bit more tricky, let me know if you have a better way to do it. 

- Fix match of undefined
It's a followup of the eslint PR that introduced the
`unnecessary-condition` rules. Unfortunately `Template` type was not
good at all.