fix(sdk): add more caching (#2111)
## Describe your changes

Fixes NAN-783

Not a fan of caching, however when running a sync it should be
deterministic and not spam our own backend for no reason. I believe we
should fetch all the data beforehand but in the meantime we can cache
more stuff.

- **Cache metadata**
It's only a part of connection, there is no reason for those to be
unsync. nb: cache is cleared after updating the metadata

- **Cache integration**
It's only used to get webhook_url however we have spike of usage (we
have no way to know if they come from us or outside the backend)