feat(webapp): [nan-839] add secondary url (#2135)
## Describe your changes
Add the ability to add a secondary webhook url. I purposely treated it
as a another column to reduce complexity and to set a hard limit of the
number of webhooks we allow customers to send to. Two should be the max
IMO.

Also fixes environment name instead of the incorrect `account`

## Issue ticket number and link
NAN-839

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is:

---------

Co-authored-by: Samuel Bodin <1637651+bodinsamuel@users.noreply.github.com>