feat: [nan-851] improve query to use the nango_config_id (#2124)
## Describe your changes
Reduce the complexity of the code since we now have the
`_nango_connections.config_id` value.

I went through the production database to ensure that value is always
populated. In some cases it was (for about 86 records) and I manually
updated the `config_id` to point to a correct value using the
`environment_id` and `provider_config_key` to match. There were two
instances where that value is NULL but those connections were created in
2023 and had no matching config

## Issue ticket number and link
NAN-851

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: