fix(temporal): handle error outside the scripts (#2163)
## Context

Fixes NAN-929

There are multiple combined errors that lead to a painful investigation:
- Temporal is wrapping all function in `activities.ts` and try/catch
them to report failure but because we added an other layer of try/catch,
all tasks are marked as successful. This prevent retry (which is a good
thing or not I don't even know anymore) but most importantly it broke
the expected return in sync.client.ts
- Sync.client.ts expected `actionHandler` to always have a shape of
`{success, error, response}` which was not the case as mentioned.
Natively Temporal is re-throwing at the place you wait for a workflow
but because of our try/catch it was just returning false (**note**: all
our workflows are broken this way but only action and webhooks are
awaited like this, I only fixed action because we are replacing
everything soon)
- The error should have at least been sent to Datadog, but because of
the foreign Temporal object it wasn't stringified properly
- The error that started this investigation `model_schema` in db being
broken somehow for this customer (it was missing `field`), I could not
reproduce, and a redeploy fixed it.

### Addition Notes
- All our workflows are somewhat broken because we have only partial
try/catch (@TBonnin you probably need to think about that with your
work)
- In a sync, if any errors happen before the activity log is created
(`activities.ts:241`) we have absolutely no trace of a sync execution
except in datadog, it should probably be the first step in any task
execution. And there are a lot of stuff that can fail.
- Error reporting is unmanageable, we have multiple telemetry.log,
errorManager, activities reporting different stuff at different place,
in `jobs` and in `server`. We should have --ideally-- one single report
in a global try/catch.
- Because we don't have a proper sentry setup, and our Datadog setup
only reports Error in span, it's simply impossible to be warned
automatically of those errors.

## Changes

- Correctly stringify the error that is sent to Datadog
- Correctly output foreign error in action 
- Fix double headers sent sync.controller
- Fix errResFromNangoErr not handling error being null
- Remove temp logging

## Test

- Throw an error in `activities.ts:runAction`
- Trigger an action

<img width="1193" alt="Screenshot 2024-05-15 at 18 55 07"
src="https://github.com/NangoHQ/nango/assets/1637651/d3cdcf5c-d7a5-4b5e-97ea-2e67e22f1e19">

<img width="1150" alt="Screenshot 2024-05-15 at 18 56 09"
src="https://github.com/NangoHQ/nango/assets/1637651/60b85703-3f69-4712-8c3f-be8251eae5b2">