feat: introduces the orchestra API and client (#2162)
## Describe your changes

Stage 2 of the orchestration v2: introducing the `orchestra` service.
The orchestra is a new service that is running the scheduler and expose
the API (via a client) to execute webhooks/actions, schedule syncs
(TBA), and fetch queued tasks (TBD)

Note: I am making heavy use of the `Endpoint` interface introduce by
Samuel a few weeks ago. It is a bit more complicated to understand but
it provides an easy way to implement type safe express endpoint. Let me
know what you think.

It is currently not wired and not running, next steps are:
- add/modify the infra to run the service in prod and staging
- add dequeuing task capabilities and task status reporting in the
orchestra client
- wire the orchestra service in the server so tasks are created for
webhook and actions (behind flag)
- wire the orchestra service in jobs so tasks are processed

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [X] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: