chore: making the scheduler db not rely on env vars (#2171)
## Describe your changes

configuring the scheduler, which is a independant library, with env vars
make it harder to understand where things are being configured and how.
We ran into issues with the records package regarding default env vars
value and behavior if/when not configured. With this commit the db url
and schema are passed into the scheduler package and we don't have a
global db object within the package. The `Tasks` function are now all
taking `db` as their first parameter, this has the nice side effect to
make it possible to run them in a transaction if needed.

## Checklist before requesting a review (skip if just adding/editing
APIs & templates)
- [ ] I added tests, otherwise the reason is: 
- [ ] I added observability, otherwise the reason is:
- [ ] I added analytics, otherwise the reason is: