fix(logs): consolidate metadata (#2187)
## Describe your changes

Contributes to NAN-908

- Add missing account, environment, config, syncConfig, connection
It's 95-99% done, there are still spots where it's partially missing,
mostly on places where we have to fetch data gradually (oauth,
controllers, etc.). To get the correct data I had to add more db calls,
even with all the refacto I tried to do, it's still not ideal to carry
partial objects everywhere, but it's getting there.

- Renamed `sync` field to `syncConfig` 
I initially used sync (as in `_nango_syncs`) but I realized in this PR
that actions do not have a row there which is confusing. So I had to
change the IDs used to fill this field, since we are not live yet it's
fine. (mapping is append compatible, the old fields are still there but
not conflicting)

- Added `providerName` field to be able to display an icon near provider
unique name

- Typed Temporal arguments correctly (needs to be handled by the new
system @TBonnin)

- Changed a few things in the UI to accommodate added/missing payload