feat(logs): UI filtering (#2193)
## Describe your changes

Fixes NAN-1011
Fixes NAN-848
Contributes to NAN-1004

- **Update schema** to accommodate filtering and aggregation
We can't aggregate on text field so we need to use index time nested
fields to have to both keyword and text. Because this schema is
incompatible and I don't want to deal with migration I just bumped the
index name. (everything will be trashed with os -> es migration anyway)

- `POST /logs/filters`
Get filters list for UI. Works more or less like the v1, except it looks
inside opensearch. That means if there is no logs you won't find the
relevant data, at first I was not convinced by this but the slow
response from the DB and the "lack" of full text search made a
compelling case to use aggregation directly from logs.

- **UI Operation filtering**
You can now filter by status, type, integration, connection and script. 

- Renamed `token` operation to `auth` 
It was hard to understand for end user and is tied to auth anyway.



## How to test?

- http://localhost:3000/dev/logs
- Trigger some action, sync, auth

## Results


https://github.com/NangoHQ/nango/assets/1637651/8cabef39-eb88-4043-a53c-cd2ee0a38020