feat(logs): opensearch -> elasticsearch (#2196)
## Describe your changes

Fixes NAN-1000

Due do lack of tooling, outdated sdk and minimal cloud support, we have
decided to use elasticsearch instead of opensearch.

- Use elasticsearch 
I haven't changed the env var to avoid creating more trouble but not
sure if it's a good idea. Let me know

- Disabled elasticsearch in self-hosted docker compose
Decision was made to have console log and empty state in the UI if this
feature is disabled. So that's it's not required in the critical path.